refactor(editor): extract useEditorHistory + editorStateModel (Phase 1b)
Move the transform state machine (editorState + undo/redo snapshot stack, with
refs mirroring state for synchronous pointer commits) into useEditorHistory, and
BASE_STATE/cloneState/stateEquals/rectEquals into editorStateModel. EditorOverlay
destructures the hook with the original names (applyState/recordState/handleUndo/
handleRedo/editorStateRef/…) so all ~40 call sites are unchanged; handleReset
stays (it also clears the message).

Behavior-preserving: golden transform + undo/redo e2e green (18/19/20 + 03/06).
